fix: bugs fixed.

+++ services/service-service.go
@@ -274,11 +274,13 @@ func initDocker(service dtos.K8sServiceDto, job structs.Job) []*structs.Command 
 	cmds := []*structs.Command{}
 	punqStructs.ExecuteBashCommandSilent("Cleanup", fmt.Sprintf("mkdir %s; rm -rf %s", tempDir, gitDir))
 	punqStructs.ExecuteBashCommandSilent("Clone", fmt.Sprintf("cd %s; git clone %s %s; cd %s; git switch %s", tempDir, service.GitRepository, gitDir, gitDir, service.GitBranch))
-	if service.App.SetupCommands != "" {
-		punqStructs.ExecuteBashCommandSilent("Run Setup Commands", fmt.Sprintf("cd %s; %s", gitDir, service.App.SetupCommands))
-	}
-	if service.App.RepositoryLink != "" {
-		punqStructs.ExecuteBashCommandSilent("Clone files from template", fmt.Sprintf("git clone %s %s/__TEMPLATE__; rm -rf %s/__TEMPLATE__/.git; cp -rf %s/__TEMPLATE__/. %s/.; rm -rf %s/__TEMPLATE__/", service.App.RepositoryLink, gitDir, gitDir, gitDir, gitDir, gitDir))
+	if service.App != nil {
+		if service.App.SetupCommands != "" {
+			punqStructs.ExecuteBashCommandSilent("Run Setup Commands", fmt.Sprintf("cd %s; %s", gitDir, service.App.SetupCommands))
+		}
+		if service.App.RepositoryLink != "" {
+			punqStructs.ExecuteBashCommandSilent("Clone files from template", fmt.Sprintf("git clone %s %s/__TEMPLATE__; rm -rf %s/__TEMPLATE__/.git; cp -rf %s/__TEMPLATE__/. %s/.; rm -rf %s/__TEMPLATE__/", service.App.RepositoryLink, gitDir, gitDir, gitDir, gitDir, gitDir))
+		}
 	}
 	if service.SettingsYaml != "" {
 		punqStructs.ExecuteBashCommandSilent("Create infrastructure YAML", fmt.Sprintf("cd %s; mkdir -p .mogenius; echo '%s' > .mogenius/%s.yaml", gitDir, service.SettingsYaml, service.GitBranch))
